requirements.txt
# Framework base de Agentes
crewai[google-genai]
crewai-tools

# Modelos y LLMs
google-generativeai

# Manipulación y estructuración de datos
pandas
pydantic

# Manejo de variables de entorno
python-dotenv

#formato
rich